requirements.txt
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
altair==5.0.1
annotated-types==0.6.0
anyio==4.6.2
arxiv==2.1.3
async-timeout==4.0.3
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
cryptography==43.0.3
dataclasses-json==0.6.5
distro==1.9.0
dnspython==2.4.2
emoji==2.14.1
etelemetry==0.3.1
eval_type_backport==0.2.2
exceptiongroup==1.2.0
feedparser==6.0.11
filelock==3.17.0
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==25.1.24
frozenlist==1.5.0
gitdb==4.0.7
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.19.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google_search_results==2.4.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
groq==0.18.0
grpcio==1.62.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.7
importlib_resources==6.5.2
isodate==0.6.1
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-google-genai==2.0.9
langchain-groq==0.2.4
langchain-mongodb==0.4.0
langchain-text-splitters==0.3.3
langdetect==1.0.9
langgraph==0.2.70
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.1.147
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==2.2.0
MarkupSafe==3.0.2
marshmallow==3.19.0
mdurl==0.1.0
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
motor==3.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numexpr==2.10.1
numpy==1.26.4
olefile==0.47
onnxruntime==1.19.2
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==10.4.0
pip==25.0
primp==0.12.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.3
prov==2.0.1
psutil==6.1.1
puremagic==1.28
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.6.1
pydeck==0.9.1
pydot==3.0.4
Pygments==2.15.1
pymongo==4.11
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==5.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-iso639==2025.1.28
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2024.1
pyxnat==1.6.3
PyYAML==6.0.2
RapidFuzz==3.12.1
rdflib==6.3.2
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.7.2
scipy==1.15.1
setuptools==75.8.0
sgmllib3k==1.0.0
simplejson==3.19.3
six==1.16.0
smmap==4.0.0
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==2.0.37
streamlit==1.42.0
sympy==1.13.3
tenacity==9.0.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.66.5
traits==7.0.2
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
unstructured-client==0.29.0
uritemplate==4.1.1
urllib3==2.3.0
watchdog==4.0.1
webencodings==0.5.1
wheel==0.45.1
win-inet-pton==1.1.0
wrapt==1.17.2
yarl==1.18.0